refactor shape class
